+++ nafas/params.py
@@ -39,8 +39,8 @@ EXIT_MESSAGE = "See you. Bye!"
 BAD_INPUT_MESSAGE = "[Error] Bad input!"
 
 CONFIG_LOAD_ERROR_MESSAGE = "[Error] Failed to load the configuration file!"
-CONFIG_INIT_ERROR_MESSAGE = "[Error] Failed to generate the configuration file!"
-CONFIG_INIT_SUCCESS_MESSAGE = "Configuration file generated successfully!"
+CONFIG_GENERATE_ERROR_MESSAGE = "[Error] Failed to generate the configuration file!"
+CONFIG_GENERATE_SUCCESS_MESSAGE = "Configuration file generated successfully!"
 
 PROGRAM_END_MESSAGE = "Well done!"