+++ test/test.py
@@ -70,10 +70,10 @@ False
 >>> result = load_config(os.path.join("test", "test_config3.json"))
 >>> result["status"]
 False
->>> result = init_config(".")
+>>> result = generate_config(".")
 >>> result
 False
->>> result = init_config("test_config4.json")
+>>> result = generate_config("test_config4.json")
 >>> result
 True
 >>> result = load_config("test_config4.json")